INT-2612: Add initial support for mongo adapters
